+++ CHANGES.rst
@@ -5,10 +5,11 @@ CHANGES
 0.6.3 (unreleased)
 ------------------
 
-- Remove unneeded test dependencies zope.app.server and
-  zope.app.component
+- Remove unneeded test dependencies zope.app.server,
+  zope.app.component, zope.app.container, and others.
 - Test on Travis CI. Only Python 2.7 is supported.
 - Update to work with zope.testbrowser 5.
+- Add PyPy support.
 
 0.6.2 (2012-06-04)
 ------------------